+++ src/index.mjs
@@ -59,13 +59,14 @@ function respondInfo(ctx) {
 }
 
 export function respondCORS() {
-  return new Response('no data collected', {
+  return new Response(null, {
     headers: {
       'Access-Control-Allow-Origin': '*',
       'Access-Control-Allow-Methods': 'GET, POST, OPTIONS',
       'Access-Control-Allow-Headers': 'Content-Type',
       'X-Frame-Options': 'DENY',
     },
+    status: 204,
   });
 }